fix(*test*.py): Make sure common ressources are given once and through gui.py because of Singleton issues
