+++ test_uexcorp_trader.py
@@ -7,42 +7,22 @@ from qasync import QEventLoop
 from gui import UexcorpTrader
 
 
-@pytest.fixture
-def app(qtbot):
-    app = QApplication.instance()
-    if app is None:
-        app = QApplication([])
-    loop = QEventLoop(app)
-    asyncio.set_event_loop(loop)
-    yield app, loop
-    loop.close()
-    app.quit()
-
-
-def test_uexcorp_trader_init(app, qtbot):
-    app, loop = app
-    trader = UexcorpTrader(app, loop)
-    qtbot.addWidget(trader)
+def test_uexcorp_trader_init(trader, qtbot):
     assert trader.windowTitle() == "UEXcorp Trader"
     assert not trader.windowIcon().isNull()
     assert trader.config_manager is not None
     assert trader.layout().count() > 0
 
 
-def test_uexcorp_trader_apply_appearance_mode(app, qtbot):
+def test_uexcorp_trader_apply_appearance_mode(trader, app, qtbot):
     app, loop = app
-    trader = UexcorpTrader(app, loop)
-    qtbot.addWidget(trader)
     trader.apply_appearance_mode("Dark")
     assert app.palette().color(trader.create_dark_palette().Window) == QColor(53, 53, 53)
     trader.apply_appearance_mode("Light")
     assert app.palette().color(trader.create_dark_palette().Window) != QColor(53, 53, 53)
 
 
-def test_tabs_exist(app, qtbot):
-    app, loop = app
-    trader = UexcorpTrader(app, loop)
-    qtbot.addWidget(trader)
+def test_tabs_exist(trader, qtbot):
     """Test that all tabs are present."""
     tabs = trader.findChild(QTabWidget)
     assert tabs is not None
